refac: rag routes
